fix: the SDK process crashes when the SA service is killed, 64-bit small devices are addpted

+++ core/common/include/softbus_server_ipc_interface_code.h
@@ -16,6 +16,8 @@
 #ifndef SOFTBUS_IPC_DEF_H
 #define SOFTBUS_IPC_DEF_H
 
+#include <stdint.h>
+
 #define SOFTBUS_SERVER_SA_ID_INNER 4700
 
 #ifdef __cplusplus
@@ -30,8 +32,8 @@ extern "C" {
 
 struct CommonScvId {
     unsigned int handle;
-    unsigned int token;
-    unsigned int cookie;
+    uintptr_t token;
+    uintptr_t cookie;
     void *ipcCtx;
     unsigned int cbId;
 };
